Feature: added identity server that will handle SSO * New IdentityService using Duende.

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -27,7 +28,6 @@ public class AuctionsController : ControllerBase
         _mapper = mapper;
         _publicEndpoint = publicEndpoint;
     }
-
     [HttpGet]
     public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string date)
     {
@@ -51,12 +51,14 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
-    [HttpPost]
+
+	[HttpPost]
+	[Authorize]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createAuctionDto);
         // TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -72,7 +74,8 @@ public class AuctionsController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
+	[Authorize]
+	public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions.Include(x => x.Item)
             .FirstOrDefaultAsync(x => x.Id == id);
@@ -80,6 +83,7 @@ public class AuctionsController : ControllerBase
         if (auction == null) return NotFound();
 
         // TODO: check seller == username
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -99,14 +103,17 @@ public class AuctionsController : ControllerBase
     }
 
     [HttpDelete("{id}")]
-    public async Task<ActionResult> DeleteAuction(Guid id)
+	[Authorize]
+	public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
-        _context.Auctions.Remove(auction);
+		// TODO: check seller == username
+		if (auction.Seller != User.Identity.Name) return Forbid();
+
+		_context.Auctions.Remove(auction);
 
         var deletedItemId = auction.Id;
         var auctionDeleted = new AuctionDeleted { Id = deletedItemId.ToString() };
